requirements.txt
rich
pyyaml
requests
sqlite-utils
